requirements.txt
fastapi
opencv-python
torch
runpod
pillow